[Refactor] #164 UserInfo 및 보험 추천 시의 user 정보들을 다루는 service 단 dto 네이밍 command로 통일
